Replace LOGIC! combinator with ISOTOPE!
LOGIC! is not a datatype, and the ~true~ and ~false~ isotope
literals can't appear in the combinator map.

This means that handling of LOGIC! must be done through a
generic isotope handler.  If this were not to be the case,
the keys in the map would have to be meta'd (or follow some
hybrid rule, like that isotopes are represented by quasiforms
and quasiforms go through a dedicated handler)