Resurrect ~end~ feature for SWITCH2 demo
The switch2 demo was based on a fanciful feature of making
a frame out of a varargs which inserts ~end~ WORD! isotopes
into the frame once arguments have been consumed.  The
partial frame would error on DO if these slots were not
filled before execution...but switch2 would fill the slots
itself with the implicit argument.  The ~end~ isotope was
known not to be an actual argument at the callsite because
frames could not be called with isotopes as non-META
parameters.

The necessity for stable isotopes to be ^META parameters was
found to be unwieldy, so the rule was changed to say that
only nihil and unstable isotopes would require ^META.  The
entire category of WORD! isotopes thus became a legal
parameter you could pass at the callsite, removing the
exemption that made the ~end~ out of band and enabled the
trick.

But due to patterns in WORD! isotopes having hardcoded
system behaviors, the idea of reserving them all and only
letting ~null~, ~true~, and ~false~ be passed in frames
came up.  This inches toward that by putting the failure on
~end~ isotopes back into the type checking phase of frame
construction, which makes the switch2 demo work again.