Try replacing ANY-CELL! with ELEMENT?
In the implementation, the name of the superset of variables
and things that are legal as array items is the "cell".

When functions like APPEND needed a name for reified items
that could be put in an array, the idea of calling these
"cells" came up (as opposed to "values", which can be held
in variables...e.g. stable isotopes).  But the incongruity
with the implementation use of the term wasn't very good.

This switches to using the term "Element" for things that
can appear in arrays, and also tries out the idea of type
specs using the ELEMENT? as a function instead of
ANY-ELEMENT! with a terminal exclamation point to feel out
the concept of letting that be the distinction between
fundamental datatypes and type constraints.