Embrace void as fetchable from plain WORD!
Due to several strange evolutions of how voids work in the
system (vanishing in non-interstitial slots, void variables
being equivalent to unset variables), there was stigma to
allowing word access to void variables.  Hence VOID had to
be a function.

Today, we can define (void: ') and access it normally.  The
embrace of this concept in PARSE allows for a way for WORD!
to give back a combinator that does not advance the input
but succeeds...(while null needs to raise an error and not
form a composable combinator at all).