Update PARSE3 and UPARSE void behaviors
This establishes void variables or evaluative results as a
synonym for empty blocks in parse--meaning they succeed without
advancing the input and synthesize a void value.

Lapses in keeping the tests working across major changes meant
several tests were still written with the concept that an
isotopic tilde was the representation of voids.  So many tests
had to be adjusted to the new concept that the evaluative
product of an apostrophe is a void, while ~ is a quasi-void
and a ~ isotope is "nihil"--a representation of an unset
variable distinct from void.