Consistency fix for UPARSE MAYBE
It may be that UPARSE's MAYBE is a bad combinator... it's
really just OPT that returns void on the parse failure it is
suppressing.  The way it was written actually tried to
MAYBE the synthesized result, which is not always possible
(e.g. when the synthesis is a none).

Ultimately the combinator might be removed, but this fixes
the behavior for now.