Trigger vaporization from empty PACK!
This revisits the idea of making voids vanish in all contexts,
giving that idea to the more fleeting construct of the empty
PACK!.  By limiting the number of constructs that make these
packs, the system becomes less risky... e.g. CASE statements
won't accidentally wind up restructured just because a void
evaluation or variable lookup appears.  But ELIDE and
COMMENT can be handled correctly.

(This breaks a lot of tests, but the fixes to those tests
will require more work.)