Swap names of NONE and NIHIL
At one point, NONE was the state of an unset variable.  This
was used as a common `return none` to indicate that a function
did not have an interesting return result (but didn't want
to be invisible).

The invention of isotopic blocks as parameter packs led to
an interesting possibility of ~[]~ isotopes...a pack that
contained no items.  This would necessarily generate an
error when you tried to assign it to a variable.  It seemed
like a natural choice for a function that didn't have an
interesting return result--so it was switched to be that
`return none` case.  The state of unset variables needed
a new name after that, so NIHIL was (awkwardly) chosen.

*But this missed a more interesting possibility*...while
void had been chosen to "vaporize" in some contexts, this
was interfering with its ability to be used as a value
when needed, or assigned to variables.  The possibilites of
dodgy vaporization were becoming more commonplace.

The higher priority of calling out unusual vaporizing
results seemed the better application of isotopic ~[]~.
Making these results rare (e.g. from narrow functions like
ELIDE and COMMENT) made more sense...and the strange name
of NIHIL seemed more suited to it.

This swaps the names back, and also does some patches that
were required to get the test suite to complete and the
build to bootstrap itself.