Unify META-GROUP! with GROUP! and GET-GROUP!
Prior to the concept of the Array_Executor(), complex
mechanics were used to determine if values were "stale" in
order to imply vaporization.

The newer blunt answer is simply that the evaluator always
writes its output into a spare cell, and copies everything
(except nihil empty isotopic packs) over into the output
after the evaluation is finished.  It's less efficient but
much easier to follow.

As a result of the simplification, complex rules which
were needed for GROUP! evaluation were no longer needed...
making it almost exactly the same as the META-GROUP!
operation (minus a "give a meta result" flag).  This does
that unification, which also means that meta groups honor
the idea that they can fully vaporize to a nihil (vs.
just produce a void in that case).