Make VOID? argument non-^META, fix typecheck
Once upon a time, voids were isotopes and the VOID? test had
to be take its argument as ^META.  But then it wasn't
tweaked to decay the value, meaning that "heavy voids" (or
any voids that were the first value in a pack) would not
pass the test.

The seemingly easy fix for this is to just take the argument
as non-^META and not worry about it.  However, that caused
some issues in the type checking, as typechecking for voids
was no longer implemented due to moving away from the
VANISHABLE flag encoding void legality.

Trying to add a proper test for voids led to difficulties
which needed to be addressed, including recursion on the
<void> argument specifier that expanded the type spec of
the value.  This fixes those issues.