Update JOIN for new datatype rules
You were able to write things like `join path! [...]` and it
would be able to tell that you were trying to join an
ANY-SEQUENCE!.  But there's no way now to test for a concrete
type passing a type constraint, because type constraints
expect an *instance* of the value.

This may suggest bringing back the idea that ANY-SEQUENCE! is
actually a block of type instances, which you FIND the types
inside of.  For now, this builds those blocks manually and
tests inside them.