Add ELIDE-IF-VOID Function
Current void/invisible behavior inherits some ideas from
early days of ~void~ BAD-WORD!s...where the vanishing state
was distinct from the reified void state.  A routine called
DEVOID would convert ~void~ into isotopic void, which would
decay to unset.

Today it's broken up more clearly.  Void is a separate state
that can't be put in blocks.  Empty parameter PACK! is the
isotopic state that always decays to variables.

ELIDE-IF-VOID is a more belabored name than DEVOID, but
clearer.  It's not yet known if this should be needed in
constructs like ANY and ALL to erase voids, but some tests
were written with the assumptions voids would vanish, and
rather than just change those tests to check for void this
threw in the construct to show it could be written if needed.