Commit to BLANK! forming as space
Interpretations of BLANK! have been progressing to where there
is even some consideration that _ should be the literal
CHAR! of space, so it's actually a SPACE?  (This would free
up BLANK? for meta-void, e.g. lone apostrophe, if needed.)

Whether things move as far as saying that blank is the literal
for space or not, it's close enough to that to say that if
blank remains its own irreducible type that it forms as space.