Eliminate UPARSE match QUOTED! from WORD!
There was the idea that an optimization of avoiding a block
by treating a lone QUOTED! fetched from a WORD! as the same
as a block rule with the quoted in it was a good idea.

This optimization doesn't make a lot of sense.  If a literal
match is intended then an @ rule can be used with more
clarity and generality.  And if the rule is truly coming
from some third party to where it's not known how many
match operations it comprises, the block is not a big cost.

Hence here that is removed, and what is legal to fetch from
a WORD! is made more explicit.  It did not seem to disrupt
anything.