Allow COPY on action isotopes
What exactly is permitted on isotopes is being felt out.  It
turns out that copying action isotopes was relied upon by a
fair amount of code, so this brings that back.

(This may point to the idea that all stable isotopes are able
to be copied--but this is starting with conservative choices
and then seeing what generalizes.)