Make || Return Nihil
There is a problem right now with LAMBDA where if the body
evaluates to nihil, you get a void back.  This is because it
uses the DELEGATE of a block to run the body, vs. a special
call of the Array_Executor() with a preload of nihil.

At the moment there are bigger fish to fry, so this just
says `|\|\||: copy :nihil` to get the behavior.