Revive WAIT and leverage libuv timers
The switch to the idea that programmers would experience
I/O written as if it were synchronous meant that WAIT on PORT!
was only applicable to server ports, which wished to wait
(essentially indefinitely) for connections to be processed.

When the EVENT! extension was excised, it inadvertently
removed that functionality.  The event loop needed to call
into libuv's event loop, but it did so in a polling manner
and relied on its own complex platform-dependent timeout
mechanics.

This revives WAIT in the network extension, but instead of
using custom per-platform code for timeouts it just uses
the libuv uv_timer_t.  Unfortunately it seems that the
uv_signal_t doesn't want to work on Windows to be able to
interrupt the loop with Ctrl-C, so this relies on a timer
that every 500 msec will check for SIG_HALT.