Fix too-few-for-multi-return cases
There is a concept that pure null is returned by functions
that want to indicate "soft failure".  This means even if
they are defined as using the multi-return output variable
form, those variables will not be proxied.  Hence something
like TRANSCODE when given an empty string will return pure
null, and not have a remainder.