Alter SPLIT-PATH to use multi-return
It was not doing so due to needing to have a /DIR refinement
in the bootstrap shim.  (Currently multi-returns appear on
the interface of a function in order to expose them to HELP,
but this should probably be done another way, e.g. with
`returns of x`)

But in the spirit of longer names being better, this works
around it by naming the output the more literate "directory".