Support QUASI-WORD! in type specs
This allows you to say things like `return: [~foo~]` to say
that you return a specfic literal isotope.

If this is decided to be a good thing, then it will extend to
let you say you return specific quoted words or other types,
which would keep you from having to define a function, e.g.

    something: func [mode ['read 'write]] [
        ...
    ]