Decay NIHIL to NONE in non-NIHIL RETURNs
You must annotate a function to indicate that it is actually
intending to return nihil and act invisibly.  But many other
cases in the system seem to toss out nihils and be the less
weird void, e.g.

    >> void? do [comment "hi"]
    == ~true~  ; isotope

So just as how things like that or WHILE loops don't want to
break the flow of evaluation by vanishing, UPARSE has a
desire to not generically permit combinators to vanish.  This
would be laborious to code into each combinator.

Here we explore the idea of just saying that functions which
try and return NIHIL but don't have it in their type signature
will return NONE instead.