Eliminate RETURN/FORWARD, add RETURN/ONLY
There was a concept that if you did not use /FORWARD, RETURN
would decay packs into single values.  This idea was based
on thinking it might be confusing if these two patterns
gave different results:

    return (...)

    ; vs.

    x: (...)
    return :x

As the notion that assignment entails implicit decay has become
more familiar, it seems better to have a DECAY native that you
use to avoid returning a pack as-is.  This keeps /FORWARD
refinements from popping up where they shouldn't be (such as
in THROW/FORWARD, so that THROW can be used with packs).

Among the consequences of this change are that the concept
that multi-return functions that wish to step out of the
multi-return paradigm and return single values like pure
NULL or VOID sometimes need another way to do it.  This is
needed since all multi-return PACKs appear to be values to
things like THEN and ELSE.  RETURN/ONLY is introduced as
a fledgling mechanic for this...but if it returns a PACK
it should likely type check it.

The commit also adds type checking for multi-returns, which
were not checked before.