Make "none" only stable get-unfriendly isotope
It's not clear just how much protection is given by forcing
people to use GET-WORD! to access stable isotopes.

Most of the time it seems to have just gotten in the way.
This removes the error until further notice, with unset
variables being the only exception for now.

The code checks incidentally for unstable isotopes.