REPLACE Enhancements
This makes REPLACE use ordinary parameters (now legal for
stable isotopes), and also removes the misguided multi-return
of the tail of the replacement.

Instead, the head and tail of each input that is replaced can
be inspected by the substitution function.  This is much more
useful in practice.

Also adds a UPARSE-based skin on top of REPLACE so that it
can mimic Red's feature of using PARSE rules in replace for
strings and binaries.