Use VOID rules in PARSE3 REWORD
Turns out that TO [] and THRU [] are broken (and have been
broken since R3-Alpha).  This was messing up REWORD which
was not based on TO VOID or THRU VOID for no-op rules.

It was easy enough to add void handling to PARSE3's TO/THRU,
vs. fixing TO [] and THRU [].