Default to not unpacking raised errors as ^META
It's more risky than with most things to overlook a raised
error when doing a meta unpack.

(Maybe meta objects too, if they can raise?)

This experimentally changes quasiforms of meta unpacks to
suppress raising an error.  It needs revisiting, but the
previous meaning (for isotopes, even stable ones) is no
longer relevant.