Fix FIND for %user.reb during boot
There was a bug where FIND was being used on a null argument
when users had a home directory with REBOL/ subdir in it.
The null needs to be converted to a VOID via MAYBE in order
to follow the "void-in-null-out" protocol, so as to make
sure opting out of the search was intended.