Disable UUID LIBRARY! Test
The code for implementing LIBRARY! is platform-dependent, so
it was kept in an extension.  LIBRARY! itself was an extension
defined datatype.

In order to make advances on the basic type system, extension
types were dropped while the issues were hammered out.  The
eventual goal is that extensions live in their own repositories
with their own issue trackers anyway, so LIBRARY extension
was moved to:

  https://github.com/metaeducation/ren-c-library/

When the time is right to bring back extension datatypes, this
will be resurrected with a priority of making it possible to
load webassembly side modules (vs. DLLs and shared libaries
on the desktop, which is not a focus).

So this disables the CI tests related to the extension.