Make network READ and WRITE errors definitional
The http server tests random disconnections, and we want to
be able to TRAP that.  However, we don't want the trapping
to use SYS.UTIL.RESCUE so that syntax errors and other
failures (like using invalid refinements) would be folded
in with the tolerance...we really just want to tolerate
problems with writing to broken pipes.

That's what definitional failures were designed to solve, so
this migrates the network based errors for READ and WRITE
to be definitional.  (There may be some subset of errors
that should not be, but for now assume they all are.)