Use ~true~ and ~false~ quasiforms in ODBC
The ODBC code used to use LOGIC! literals #[true] and #[false]
to represent logic in blocks.

Now that there is no separate "logic type", the isotopic words
for true and false are used.  But to put them in a block, they
have to be their QUASI! forms as ~true~ and ~false~.

(This is a good enough change to get a round trip on the
parameter blocks, but may need improved ergonomics for a
real data extraction use.)