Change from MAYBE/OPT to TRY in PARSE
After consideration of the tradeoffs, TRY has been picked as
a more literate English-like choice for the function that OPT
has historically had in PARSE.

MAYBE had been tried, but it is longer and contentious with
the behavior of MAYBE in non-PARSE scenarios.  TRY is also
consistent with Haskell Parsec terminology for when a
parser will rewind and continue.