Fix MAYBE definition in bootstrap shim
MAYBE once had a different meaning, which was that it would
not update a variable in an assignment if the right hand
side was null.

   >> x: 10

   >> x: maybe if false [<a>]
   == 10

   >> x
   == 10

There is no current choice for what such a construct would
be called.  But MAYBE was retaken to mean NULL-TO-VOID.

This adjusts the definition and callsites appropriately in
the bootstrap shim.