Disallow testing for IS_TRASH in release builds
The idea of a "trash" cell is that debug builds can protect
against reading such cells before they are filled with
meaningful data.  But unlike an ordinary uninitialized cell,
these trash cells are ignored by the garbage collector...so
they can sit in arrays spanning evaluations.

There was a misguided idea that IS_TRASH() in a release build
was defined to false, based on the concept that no such cells
would exist.  This broke a case in the APPLY implementation
where the trash state was being used to carry meaning...because
in the release build initializing a cell to trash and then
asking if it was trash would return false.  The correct way
to avoid this problem is to not define IS_TRASH() in the
release build to catch such incorrect uses.

APPLY was changed to use a NONE state for the signal, as it
needed a signal that could be read as well as written.