Simplify BLANK! and VOID, update comments
At one point there was an idea that since NULL did not have
a "payload", its cell contents could be used to store some
information about the file and line where it was produced,
even in release builds without making the cells larger.
(DEBUG_TRACK_EXTEND_CELLS makes all cells larger so that
any type stores some tracking information, but this adds
significant cost so is not exposed to users as a feature.)

The concept of storing tracking info in null cells seemed
like it might come in handy if exposed to users, particularly
since nulls would often be a trigger for errors.  Other
types that didn't need payloads--like BLANK! and VOID--were
less hypothetically useful for release builds to know where
they were produced, but they were given the capability
anyway via a generalized "Init_Nothing()" initializer.

Design evolution made nulls need a payload, as they were
actually isotopic words--holding the symbol for "null" as
well as an optional binding.  That meant the feature wound
up only applying to blanks and voids.

Given that this feature was only hypothetical and never
exposed--as well as no longer applicable to nulls--this gets
rid of the concept.

The opportunity is also taken to do some work on updating
the comments to reflect the current state of the system as
it pertains to null/void/blank