Handle raised errors in console code
The PRINT-RESULT function was not annotated to say that it
is able to handle raised error isotopes for its meta parameter.
Hence the console skin would choke on simple inputs like:

    >> raise "test"

This adds the capacity for taking raised errors to PRINT-RESULT
and also changes the name of the tag from `<fail>` to `<raised>`.