Give TAKE definitional error if nothing to take
TAKE is an example of a function with a side effect of
modifying series, that has historically had a bit of a blind
spot in that TAKE of an empty series would not error.

Rebol2/R3-Alpha/Red were particularly problematic because
the returned value of #[none] could be mistook for taking a
block that contained #[none].

    rebol2>> take []
    == #[none]

    rebol2>> take [#[none]]
    == #[none]

Ren-C's introduction of "definitional errors" means that
it's possible for TAKE to RAISE an error that can be trapped
cleanly, knowing that it originated from the TAKE.

    >> take []
    ** Error: Can't TAKE, no value available (consider TRY TAKE)

    >> try take []
    == ~null~  ; isotope

This makes the change, although it leaves the historical
behavior of TAKE/PART to always return a series (hence the
failure mode gives back empty series instead of null).  That
probably needs review.