Fix ABOUT command
The ABOUT command used a PARSE-based function MAKE-BANNER,
which relied on PARSE, but had no check to make sure the
parse rules reached the end of the input.

There had been a longstanding unknown issue that ABOUT was not
working, because the parse rule expected PATH! in slots that
were overtaken by TUPLE! in Ren-C.

This issue was caught by making PARSE3 error by default when
it couldn't reach the end of input (subsequent commit).