INHERIT-META reacts to NONE, not VOID
When making inherited metadata about a function was done, it
originally reacted to either VOID or NULL.  The meaning of
VOID changed and the relevant thing to react to is NONE (e.g.
an isotopic void).

(All of this should be rethought, but this just gets things
like HELP FUNCTION working.)