Various fixes to HELP functions
There are a lot of open issues in the design of HELP, based on
non-trivial questions about how meaningful the runtime
representation of source can be reflected.  So it has gone
through some atrophy.

However, the tests should not raise errors.  This patches the
code to make the %help.test.reb at least not fail, even if the
information isn't always useful at this moment.