Make PARSE3 RAISE errors on unsuccessful parse
The design choice to raise a definitional error on incomplete
parses worked so well for UPARSE that this goes ahead and
patches the behavior into PARSE3.

As a result of making this change, bugs were discovered and
unclear code was rewritten to make it obvious when accepting
incomplete states was intentional.