No PARSE3 END, upgrade command-line PARSE rule
PARSE3 hadn't yet deprecated the plain word END, which is done
via the tag `<end>` in the UPARSE model (leaving END available
for a variable name).

This does that deprecation, which drew attention to the odd
usage in parsing command-line arguments.  The very old code
was based on looping over the individual items in argv, and
parsing them as strings--which meant it was using string
parsing vs. block parsing.  Hence matching a complete string
was not possible without writing e.g. `["--version" <end>]`

So this moves the iteration code into the parse itself, so
that it becomes a BLOCK! parse, and strings can be matched
in whole.  Cases which want to get more detailed enter
string matching via INTO (which should become SUBPARSE).