Bootstrap fixes for Haiku commit, haiku-32 0.5.4
Bootstrapping with old executables requires rather stylized
coding (no SWITCH/TYPE or NULL! type).  This adjusts for that.

Additionally, the "humorous" choice to number Haiku as 0.5.75
is replaced with a more standard numbering of Haiku 32-bit as
0.5.4 and 64-bit as 0.5.40

Adds comments to the patches in %sys-zlib.h and %fs.c to the
effect of that these should be submitted to zlib and libuv,
as they would be lost on updates of those sources from their
respected repository.