Simplify refinement and return typechecking
When refinements were adapted to be special cases of paths,
and characters as special cases of issues, this introduced
the need to be able to check not just a fundamental type but
a "type constraint".  This was done very hackily for just
those examples initially, but the system is now migrating
to where arbitrary functions can be used in type blocks as
constraints.

The messy evolution hadn't really been cleaned up (including
a detour into the type constraint of tuples with leading
blanks as being "predicates").  There was also separate type
checking for refinements which hadn't been folded in fully
when refinements became "their own arguments", that had some
code in specialization that was more complex than it needed
to be.

This takes a step toward a bit more clarity, deleting unneeded
bits and fusing the refinement type checking with the parameter
checking.  Also separates out the return type checking into
its own function for clarity.