Simplify Argument Coercion
Prior to type checking being done by means of calling typecheck
functions, there had to be special tags in type specs for things
that couldn't be expressed as a "DATATYPE!".  This got hairy
as things like unstable isotopes were introduced, because tags
were needed to express things like "this meta parameter is
actually interested in raised errors, not packs" or "this
meta parameter is actually interested in packs, not raised
errors."

This removes the tags `<nihil>`, `<raised>`, and `<pack>` in
favor of NIHIL?, RAISED?, and PACK? generic functions.  Then in
argument type checking, there is explicit coercion for unstable
isotopes were the isotopes are passed as-is through the type
checking process first, then if that does not succeed they
decay and are passed through again.