Don't classify output parameters as refinements
Prior to the idea of PACK! for multi-return, there was an idea of
being able to pass in variables via refinement as a feature of
multi-returns.  This had bad properties which motivated the
current model.  However, output parameters were still categorized
as refinements.  This led to a bug where not supplying any type
spec to an output parameter meant it was typechecked like a
no-argument refinement, and could only be null or a blackhole.

This also does an adjustment so that outputs do not come back
from PARAMETERS OF for a function, but rather OUTPUTS OF.  This
required changes to HELP.