Introduce intrinsics
This makes a new kind of function which exposes its implementation
to internal code as a C function which takes a single argument and
an output cell.  The application of this is to be able to make
fast calls to that function without building a frame.

The way this is done is to fold intrinsics in with ordinary
natives, where there's still a single C function associated with
the native--but it's not the dispatcher.  Instead, the C function
is associated with the native, but with a common dispatcher that
simply plucks the C function out and calls it with the first
argument from the frame (and the output cell).

Where this provides value is if internal code sees one of these
functions, it can bypass the generic calling mechanisms and
extract the simpler C function out and use that.