Get rid of JUST
JUST was a means of generating a literal quoted value, like QUOTE
added on top of THE:

   >> just a
   == 'a

The initial JUST-ification for this idea originated in the time when
quote marks would suppress splicing in append:

   >> append [a b c] first [[d e]]
   == [a b c d e]

   >> append [a b c] first ['[d e]]
   == [a b c [d e]]

When this was true, it was thought that an operator which folded
together literal arguments with quoting would be useful:

   >> append [a b c] just d
   == [a b c d]

   >> append [a b c] just [d e]
   == [a b c [d e]]

Ultimately, Ren-C shifted away from the concept of adding a quote
level to suppress splicing, to *subtracting* a quote level to
request splicing...with isotopes representing "values at a quoting
level of -1".  This changed the landscape completely.

Since JUST had no obvious application after that change, it doesn't
really need to hang around...and the word can be reclaimed for
something else.