Handle phase errors more elegantly
A noble idea for reporting type checking errors that originate
from internal phases was implemented in a way that made it hard
to pass in symbols to argument type errors.  This refactors the
code and uses a switcheroo in the error ID to accomplish sharing
of code between the internal phase error and the ordinary arg
type error.