Add BASIC-READ and BASIC-WRITE for file I/O in WASI
The WASI system call layer for files doesn't provide everything
that the Posix implementation of libuv requires.  However it is
able to implement the basics of C's fread() and fwrite() for the
sandboxed file system access you give it.

There should probably be an extension for basic file I/O which
can be used on systems that don't support libuv, as a kind of
lowest common denominator.  But for now this just pokes a couple
of routines into the executable when built as libuv to serve
as a simple test.