Make date comparison more lax
An attempt to be more rigorous made it so that comparisons of dates
with time and zone components could not be done with dates that
did not have them.  It raised an error.  This turned out to not be
all that pragmatic, and it's better to have comparisons do some
lightly intuitive guessing in non-strict mode.  For instance, to
say that (26-Jul-2021/7:41:45.314 > 26-Jul-2021) is false, because
they both lie on the same day.