Type signature for wrapper combinator returns pack
A pending change where typechecking ANY-VALUE! does not cover
non-decayable packs (such as NIHIL? or those whose elements
contain packs or errors) exposed that the wrapper inside of
UPARSE needed to say it returned PACK?.

This also raised attention to the fact that the COMBINATOR?
test had been broken, when WORDS-OF stopped giving back
return types.  You can't determine if something is a combinator
based on returns due to the fact that wrapped combinators
will not appear to have multi-returns.

Fixes an error message covering the above case while doing
this (attempt to form a BLOCK!, needed an UNSPACED)