Rethink TYPECHECKER as using the Intrinsic mechanism
The first attempt to speed typechecking was to notice usages
of functions like INTEGER? or ANY-VALUE? in type specs, and
avoid calling those as functions...but instead just make an
assumption on what those functions meant in their implementation.
Hence `Datatype/Typeset_Typechecker_Dispatcher()` were looked for
as what the function looked up by words like INTEGER? or ANY-VALUE?
were using.

Then there was the idea that some functions that weren't based on
fundamental types or premade typesets could be optimized as well.
These were optimized not just in typecheck blocks, but also in the
evaluator...so if the calls were made in ordinary code those would
be accelerated as well.  This mechanism was named "intrinsics".

By adding a parameter to the intrinsics which provides a pointer to
the function itself, it meant the intrinsic could be sensitive to
some additional information that was associated with the function
at creation time.  This allowed the typechecker functions to become
intrinsics, since implementations for things like INTEGER? and
BLOCK? could both use the Instrinsic_Dispatcher() but get a variance
in behavior based on receiving the action parameter and looking in
it for what the instance was supposed to check.