Fix Swap_Series_Content() for dynamic series
At one time, the SER_USED() byte of 255 would convey the state
that a series was dynamically allocated.  This was changed so
that SERIES_FLAG_DYNAMIC was used instead.  However, the swap
series content function was still assuming it only needed to
swap the SER_USED() byte and that would transfer the dynamic
status.  It needs to proxy the flag, now that it's a flag.