RULE-START and RULE-END for PARSE, also /HOOK
This makes a long-planned step away from the /VERBOSE option
to a generalized parse hook, that is called on each instance
of a "combinated" parser.

Fields for a rule's start and end are added, so that the hook
knows the span of source code from which the parser was made.