DATE! related comparison updates
Lacking any great answer, date comparisons were made more lax
with non-strict equality, so that if a less-specific date (e.g.
missing a time) was compared with a date that had a time, it
would be equal if the date part was the same.  The tests did
not get updated, in part due to an oversight in the test driver
regarding tests that don't raise expected failures not reporting
the lack of a failure if they succeed.