PARSE-FURTHEST stopgap
Transitioning to raising an error on incomplete parses broke
the idea of making "how far did you get" an ordinary return
value (as did making it so that combinators could themselves
return multi-values, hence making the overall PARSE result
a multi-return).

This works in a hacky hooked PARSE called PARSE-FURTHEST
which takes a variable to update.  It's a placeholder to keep
the functionality working until a better way to expose it
comes up.