Exempt null from isotopic word exemption in RAISE
The visual "error-ness" of (fail ~unreachable~) led to it being
used instead of (fail "unreachable").  This may not be viable
with rules about whether users can instantiate and use arbitrary
isotopic words... but beyond that, this was disrupting the
interpretation of (fail) with no arguments because that made the
argument a null due to being a non-^META `<end>` parameter.